hall_booking_app/functions/requirements.txt
# Firebase SDKs
firebase-functions
firebase-admin

# Web framework (Flask) and related libraries
flask
flask-cors
pytz
python-dateutil

